before the end

+++ src/main/java/com/example/frequencies/Users.java
@@ -16,6 +16,9 @@ import javafx.stage.StageStyle;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.*;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -25,19 +28,19 @@ public class Users implements Initializable {
     Button profileBouton, usersBouton, abonnesBouton, frequencesBouton, deconnexionBouton, ajouterBouton, deleteBouton, refreshBouton, editBouton;
     @FXML
     TableView <User> users;
+    Timestamp timestamp = new Timestamp(new Date().getTime());
     @FXML
     TableColumn <User, String> idColumn, prenomColumn, nomColumn, matColumn, sexeColumn, phoneColumn, emailColumn, adressColumn, mdpColumn;
     ObservableList<User> Users = FXCollections.observableArrayList();
+    ArrayList<User> arrayUsers = new ArrayList<>();
     Main main = new Main();
     DataBaseConnection dbConnection = new DataBaseConnection();
     Connection connection = dbConnection.getConnection();
-    ResultSet resultSet;
-    PreparedStatement statement;
+    ResultSet resultSet, resultSet1;
+    PreparedStatement statement, statement1;
     User user;
     AjoutUser ajoutUser;
-    String query;
-
-
+    String query, query1;
 
         public void profile(ActionEvent event) throws IOException {
             main.changeScene("accueil.fxml");
@@ -63,7 +66,7 @@ public class Users implements Initializable {
         public void modifier(ActionEvent event) throws IOException{
             user = users.getSelectionModel().getSelectedItem();
             FXMLLoader fxmlLoader = new FXMLLoader();
-            fxmlLoader.setLocation(getClass().getResource("ajoutUser.fxml"));
+            fxmlLoader.setLocation(getClass().getResource("ajoutUtilisateur.fxml"));
             try {
                 fxmlLoader.load();
             } catch (IOException e){
@@ -71,8 +74,8 @@ public class Users implements Initializable {
             }
             ajoutUser = fxmlLoader.getController();
             ajoutUser.setUpdate(true);
-            //ajoutUser.setTextFields(user.getId(), user.getNom(), user.getPrenom(), user.getEmail(), user.getAdresse(),
-                    //user.getTelephone(), user.getMdp(), user.getMatricule(), user.getSexe());
+            ajoutUser.setTextFields(user.getId(), user.getNom(), user.getPrenom(), user.getEmail(), user.getAdresse(),
+                    user.getTelephone(), user.getMdp(), user.getMatricule(), user.getSexe());
 
             Parent parent = fxmlLoader.getRoot();
             Stage stage = new Stage();
@@ -97,16 +100,53 @@ public class Users implements Initializable {
                         resultSet.getString("adresse"),
                         resultSet.getString("mdp")));
                 users.setItems(Users);
-
              }
-
         }
-        public void supprimer(ActionEvent event) throws IOException, SQLException {
+        public void supprimer(ActionEvent event) throws SQLException {
+
+            connection.setAutoCommit(false);
             user = users.getSelectionModel().getSelectedItem();
-            query = "DELETE FROM `users` WHERE `users`.`id` = " + user.getId();
-            statement = connection.prepareStatement(query);
-            statement.execute();
-            refresh();
+
+            Alert alert, alert1;
+            alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Confirmation");
+            alert.setHeaderText("Suppression de fréquence.");
+            alert.setContentText("Etes-vous sûrs de vouloir supprimer cet utilisateur?");
+            Optional<ButtonType> result = alert.showAndWait();
+
+            if (result.get() == ButtonType.OK) {
+
+                query = "DELETE FROM `users` WHERE `users`.`id` = " + user.getId();
+                query1 = "INSERT INTO `operation` (`type`, `date`, `userConnected`, `idAbonne`, `idFrequence`, `idUser`) VALUE(?, ?, ?, ?, ?, ?)";
+
+                statement = connection.prepareStatement(query);
+                statement1 = connection.prepareStatement(query1);
+                statement1.setString(1, "suppression");
+                statement1.setTimestamp(2, timestamp);
+                statement1.setInt(3, Login.userOnline.getId());
+                statement1.setInt(4, 0);
+                statement1.setInt(5, 0);
+                statement1.setInt(6, user.getId());
+                try {
+                    statement1.execute();
+                    statement.execute();
+                    connection.commit();
+                    refresh();
+                    alert1 = new Alert(Alert.AlertType.INFORMATION);
+                    alert1.setTitle("Succès");
+                    alert1.setHeaderText("Succès!");
+                    alert1.setContentText("L'utilisateur a été supprimé avec succès.");
+                    Optional<ButtonType> result1 = alert1.showAndWait();
+                    if (result1.get() == ButtonType.OK){
+                        alert1.close();
+                    }
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+                refresh();
+            } else {
+                alert.close();
+            }
         }
 
     @Override
